Move clipboard access to service
